Refatorando requisitos1, 2, 4 e adiciona requisito 3
