+++ src/components/ShoppingCartIcon.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+class ShoppingCartIcon extends React.Component {
+  render() {
+    return (
+      <img src="https://image.flaticon.com/icons/png/512/263/263142.png" alt="ícone de carrinho de compras" />
+    );
+  }
+}
+
+export default ShoppingCartIcon;
